1.add log 2.Message length limit 2048 3.Parameter conversion float64 -> int

+++ open_im_sdk/ws_wrapper/ws_local_server/ws_conversation_msg.go
@@ -140,7 +140,7 @@ func (wsRouter *WsFuncRouter) GetOneConversation(input string, operationID strin
 	}
 
 	userWorker := open_im_sdk.GetUserWorker(wsRouter.uId)
-	userWorker.GetOneConversation(m["sourceID"].(string), m["sessionType"].(int), &BaseSuccFailed{runFuncName(), operationID, wsRouter.uId})
+	userWorker.GetOneConversation(m["sourceID"].(string), int(m["sessionType"].(float64)), &BaseSuccFailed{runFuncName(), operationID, wsRouter.uId})
 }
 
 func (wsRouter *WsFuncRouter) GetMultipleConversation(conversationIDList string, operationID string) {
@@ -256,7 +256,7 @@ func (wsRouter *WsFuncRouter) CreateVideoMessageFromFullPath(input string, opera
 	if !wsRouter.checkKeysIn(input, operationID, runFuncName(), m, "videoFullPath", "videoType", "duration", "snapshotFullPath") {
 		return
 	}
-	msg := userWorker.CreateVideoMessageFromFullPath(m["videoFullPath"].(string), m["videoType"].(string), m["duration"].(int64), m["snapshotFullPath"].(string))
+	msg := userWorker.CreateVideoMessageFromFullPath(m["videoFullPath"].(string), m["videoType"].(string), int64(m["duration"].(float64)), m["snapshotFullPath"].(string))
 	wsRouter.GlobalSendMessage(EventData{cleanUpfuncName(runFuncName()), 0, "", msg, operationID})
 }
 
@@ -286,7 +286,7 @@ func (wsRouter *WsFuncRouter) CreateSoundMessageFromFullPath(input string, opera
 	if !wsRouter.checkKeysIn(input, operationID, runFuncName(), m, "soundPath", "duration") {
 		return
 	}
-	msg := userWorker.CreateSoundMessageFromFullPath(m["soundPath"].(string), m["duration"].(int64))
+	msg := userWorker.CreateSoundMessageFromFullPath(m["soundPath"].(string), int64(m["duration"].(float64)))
 	wsRouter.GlobalSendMessage(EventData{cleanUpfuncName(runFuncName()), 0, "", msg, operationID})
 }
